Moved icons over to ts-components
