+++ packages/ts-components/src/assets/TooltipIcon.tsx
