+++ packages/ts-components/src/assets/icons.stories.mdx
@@ -1,0 +1,1 @@
+overrides={{size: "sizing050"}}
