+++ packages/ts-components/src/assets/index.tsx
@@ -1,0 +1,1 @@
+export { default as TooltipIcon } from '../assets/TooltipIcon';
