+++ packages/ts-components/src/components/job-title/job-title.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TooltipIcon } from '@times-components/ts-newskit';
+import { TooltipIcon } from '../../assets';
 import {
   CloseButton,
   IconButton,
