+++ packages/ts-newskit/src/index.ts
@@ -6,6 +6,5 @@ export {
   UpdateButtonWithDelay
 } from './components/misc/update-button/update-button-with-delay';
 export { Banner } from './components/misc/banner/banner';
-export { TooltipIcon } from './assets';
 
 // TODO Remove after tests - borisovy
